requirements.txt
# Core dependencies
beautifulsoup4==4.12.3
requests==2.31.0
newspaper3k==0.2.8
lxml==5.1.0
html5lib==1.1

# AI and ML
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.8
chromadb==0.4.22

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
click==8.1.7 